requirements.txt
pytest==7.4.3
PyYAML==6.0.1
coinbase-advanced-py>=1.2.0
httpx
uvicorn
fastapi
prometheus_fastapi_instrumentator
kafka>=1.3.5